server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "server for docquer"
authors = ["Abhiram86 <abhiramalla86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.115.3"
langchain = "^0.3.4"
python-dotenv = "^1.0.1"
pymongo = "^4.10.1"
uvicorn = "^0.32.0"
pypdf2 = "^3.0.1"
python-docx = "^1.1.2"
langchain-text-splitters = "^0.3.1"
pinecone = "^5.3.1"
sentence-transformers = "^3.2.1"
easyocr = "^1.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
